Legacy `ICARUSChannelMap` service split
The new layout is:
 * service provider and service interface (`ICARUSChannelMap.h`) in the library
 * implementation of the service in `ICARUSChannelMap.cc`
 * registration of the service in `ICARUSChannelMap_service.cc`
